More work on GenCPU backend test (currently not passing)

+++ genericcpu-backend/cmake/src/matmul.h
@@ -7,17 +7,17 @@ extern "C" {
 
 #endif
 JNIEXPORT void JNICALL Java_me_mikex86_scicore_backend_impl_genericcpu_jni_MatmulJNI_matmul(JNIEnv *, jclass, jint transa, jint transb,
-                                                                                            jlong m, jlong n, jlong k,
+                                                                                            jint m, jint n, jint k,
                                                                                             jlong alphaPtr,
                                                                                             jlong aPtr,
                                                                                             jint aType,
-                                                                                            jlong lda,
+                                                                                            jint lda,
                                                                                             jlong betaPtr, jlong bPtr,
                                                                                             jint bType,
-                                                                                            jlong ldb,
+                                                                                            jint ldb,
                                                                                             jlong cPtr,
                                                                                             jint cType,
-                                                                                            jlong ldc);
+                                                                                            jint ldc);
 
 #ifdef __cplusplus
 }
